remove allow dead_code annotations
