+++ lgn-provers/src/provers/v1/query/mod.rs
@@ -2,12 +2,14 @@ use crate::provers::v1::query::prover::StorageQueryProver;
 use crate::provers::v1::query::task::Querying;
 use tracing::info;
 
-pub(crate) mod euclid_prover;
 pub(crate) mod prover;
 pub mod task;
 
 #[cfg(feature = "dummy-prover")]
 pub(crate) mod dummy_prover;
+
+#[cfg(not(feature = "dummy-prover"))]
+pub(crate) mod euclid_prover;
 
 pub const MAX_NUM_RESULT_OPS: usize = 20;
 pub const MAX_NUM_RESULTS: usize = 10;
